murmur/pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "murmur"
dynamic = ["version"]
requires-python = ">=3.11"
dependencies = [
    "faster-whisper",
    "sounddevice",
    "numpy",
    "pynput",
    "pystray",
    "Pillow",
    "platformdirs",
    "pyautogui; sys_platform == 'win32'",
    "pyperclip; sys_platform == 'win32'",
    "pywin32; sys_platform == 'win32'",
]

[project.optional-dependencies]
wakeword = ["openwakeword"]
build = ["pyinstaller"]
test = ["pytest", "ruff"]
dev = ["bump-my-version", "ruff"]

[project.scripts]
murmur = "murmur.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "murmur.__version__"}

[tool.bumpversion]
current_version = "0.1.0"
commit = true
tag = true
tag_name = "v{new_version}"

[[tool.bumpversion.files]]
filename = "src/murmur/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
